+++ Scripts/Player/PlayerStats.cs
@@ -4,7 +4,7 @@ public class PlayerStats : Entity
 {
     public float attackRange = 2f;
 
-    public PlayerStats(float _health = 1f, float _attackDamage = 1f, float _speed = 1f, int _age = 20)
+    public PlayerStats(float _health = 1f, float _attackDamage = 1f, float _speed = 1f, float _age = 20)
     {
         health = _health;
         attackDamage = _attackDamage;
